# swagger.yaml
openapi: 3.0.3
info:
  title: Werfie API
  description: |
    Complete API documentation for Werfie (X-Clone) platform.
    - **Client Gateway**: http://localhost:3001
    - **Admin Backend**: http://localhost:3012
    - **Messaging Service**: http://localhost:3019 (direct or via gateway)

    ## Quick test (always 200)
    **GET /api/health** - No auth. Returns 200 when gateway is up.

    ## Auth Flow (to get 200 on protected endpoints)
    1. Call **POST /api/auth/login** with `{"email":"user1@xclone.com","password":"password123"}`
    2. Copy the `accessToken` from the response (from `data.accessToken`)
    3. Click **Authorize** (lock icon, top right), paste ONLY the token (no "Bearer " prefix)
    4. Protected endpoints (/api/auth/me, /api/auth/logout, /api/auth/change-password) will now return 200

    ## Proxied endpoints (Users, Posts, etc.)
    Endpoints like /api/users/*, /api/posts/* are proxied to microservices. Ensure those services are running (see ecosystem.config.js) or you may get 5xx.
  version: 1.0.0
servers:
  - url: /
    description: Same origin (use this - no CORS)
  - url: http://localhost:3001
    description: Client Gateway (Auth, Users, Posts, etc.)
  - url: http://127.0.0.1:3001
    description: Client Gateway (127.0.0.1)
  - url: http://localhost:3019
    description: Messaging Service (direct)
  - url: http://localhost:3012
    description: Admin Backend
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste the accessToken from /api/auth/login (no "Bearer " prefix)
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...
paths:
...
  /api/business:
    get:
      summary: Get business profile
      tags: [Business]
      responses:
        '200': { description: Business profile }
    post:
      summary: Create/register business
      tags: [Business]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                category: { type: string }
                description: { type: string }
      responses:
        '201': { description: Created }
...
  /api/business/boost:
    post:
      summary: Boost post
      tags: [Business]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                postId: { type: string }
      responses:
        '200': { description: Boosted }
  /api/business/team:
    get:
      summary: Get team
      tags: [Business]
      responses:
        '200': { description: Team members }
    post:
      summary: Add team member
      tags: [Business]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                memberId: { type: string }
                role: { type: string }
      responses:
        '200': { description: Added }
  /api/business/team/{memberId}:
    patch:
      summary: Update team member
      tags: [Business]
      parameters:
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      summary: Remove team member
      tags: [Business]
      parameters:
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Removed }
  /api/business/verify-domain:
    post:
      summary: Verify domain
      tags: [Business]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domain: { type: string }
      responses:
        '200': { description: Verification status }

  # ============ INSTITUTIONAL ============
...